requirements.txt
xlrd
nltk
pandas
matplotlib
seaborn
pyspellchecker
gensim